requirements.txt
iso8601>=0.1.8
lxml>=3.2.1
pyparsing
pyramid>=1.4.2
pyramid_tm>=0.7
pyramid_zcml>=1.0.0
python-graph-core
sqlalchemy>=0.8.1
zope.sqlalchemy>=0.7.2
